Komponent - wybór aktualnego dnia z tygodnia. Zwraca aktualną datę
